requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyGithub
httpx
pytest
pytest-playwright
playwright
jsonschema
